.rb files generated for Slack commands / renaming get_ling.rb -> say_hello.rb
