+++ mb_ruby.rb
@@ -1,3 +1,6 @@
 require 'slack-ruby-bot'
-require 'mb_ruby/commands/get_link'
+require 'mb_ruby/commands/insult_member_team'
+require 'mb_ruby/commands/screenshot_page'
+require 'mb_ruby/commands/wolfram_research'
+require 'mb_ruby/commands/say_hello'
 require 'mb_ruby/bot'
