+++ mb_ruby/bot.rb
@@ -6,12 +6,12 @@ module MbRuby
   class Bot < SlackRubyBot::Bot    
     help do
       title 'My Ruby Bot'
-      desc 'This bot is about getting a given youtube video'
+      desc 'This office bot, which will heplp you as an office assistant'
       
-      command :get_link do
-        title 'get youtube video link'
-        desc 'Returns the url of a given youtube video topic'
-        long_desc 'Returns the url of a given youtube video topic'
+      command :say_hello do
+        title 'say_hello'
+        desc "Welcome command, to recieve you properly"
+        long_desc "Welcome command, to recieve you properly"
       end
     end
   end
